+++ src/main/java/oripa/gui/bind/state/LocalPaintBoundStateFactory.java
@@ -7,7 +7,7 @@ import oripa.appstate.StateManager;
 import oripa.gui.bind.state.action.PaintActionSetterFactory;
 import oripa.gui.presenter.creasepattern.EditMode;
 import oripa.gui.presenter.creasepattern.GraphicMouseAction;
-import oripa.gui.viewsetting.ChangeViewSetting;
+import oripa.gui.viewchange.ChangeViewSetting;
 
 /**
  * Helper class
