+++ src/main/java/oripa/gui/presenter/main/ArrayCopyDialogPresenter.java
@@ -22,7 +22,7 @@ import oripa.domain.paint.ArrayCopyParameter;
 import oripa.domain.paint.PaintContext;
 import oripa.domain.paint.arraycopy.ArrayCopyCommand;
 import oripa.gui.view.main.ArrayCopyDialogView;
-import oripa.gui.viewsetting.ViewScreenUpdater;
+import oripa.gui.view.main.ViewScreenUpdater;
 import oripa.util.Command;
 
 /**
