+++ src/main/java/oripa/gui/presenter/main/MainFramePresenter.java
@@ -49,9 +49,9 @@ import oripa.gui.view.main.MainFrameDialogFactory;
 import oripa.gui.view.main.MainFrameView;
 import oripa.gui.view.main.MainViewSetting;
 import oripa.gui.view.main.PainterScreenSetting;
+import oripa.gui.view.main.ViewScreenUpdater;
+import oripa.gui.view.main.ViewUpdateSupport;
 import oripa.gui.view.util.ChildFrameManager;
-import oripa.gui.viewsetting.ViewScreenUpdater;
-import oripa.gui.viewsetting.ViewUpdateSupport;
 import oripa.persistence.dao.AbstractFilterSelector;
 import oripa.persistence.doc.CreasePatternFileTypeKey;
 import oripa.persistence.doc.DocFilterSelector;
