+++ src/main/java/oripa/gui/view/main/ViewScreenUpdater.java
@@ -1,4 +1,4 @@
-package oripa.gui.viewsetting;
+package oripa.gui.view.main;
 
 public interface ViewScreenUpdater {
