+++ src/main/java/oripa/gui/viewsetting/main/PainterScreenSettingImpl.java
@@ -6,6 +6,7 @@ import java.beans.PropertyChangeSupport;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import oripa.gui.view.main.InitialVisibilities;
 import oripa.gui.view.main.PainterScreenSetting;
 
 public class PainterScreenSettingImpl implements PainterScreenSetting {
